Change type from double to decimal for Attack and Health Hero attributes

+++ rpgsaga/src/RpgSaga.DLCs/Rogue/Heroes/Knight.cs
@@ -4,12 +4,12 @@ namespace RpgSaga.DLCs.Rogue.Heroes;
 
 public sealed class Knight : Hero
 {
-    public Knight(string firstName, string lastName, double health, double attack) 
+    public Knight(string firstName, string lastName, decimal health, decimal attack) 
         : base(firstName, lastName, health, attack)
     {
     }
 
-    public static Knight Create(string firstName, string lastName, double health, double attack)
+    public static Knight Create(string firstName, string lastName, decimal health, decimal attack)
     {
         return new Knight(firstName, lastName, health, attack);
     }
